% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Calculates age}
\usage{
age(dob, age.day = today(), units = "years", floor = TRUE)
}
\arguments{
\item{dob}{A date object, generally the date of birth}

\item{age.day}{A date object, the date for which you want the age}

\item{units}{Time unit, see unites from lubridate::duration}

\item{floor}{Logical, returns integer}
}
\value{
Age of observation in defined units, years in default
}
\description{
Calculates age
}
