% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_to_indicator.R
\name{character_to_indicator}
\alias{character_to_indicator}
\title{Transforms a character column into indicator columns}
\usage{
character_to_indicator(in_data, in_unique_identifier, in_variable, in_prefix)
}
\arguments{
\item{in_data}{The dataframe with the character column}

\item{in_unique_identifier}{Unique identifiers of the dataframe. This can except a vector.}

\item{in_variable}{Character column in the dataframe that needs to be transformed}

\item{in_prefix}{Prefix of the new indicator columns created}
}
\value{
Dataframe with unique identifier(s) and indicator columns
}
\description{
Transforms a character column into indicator columns
}
