% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders_oes_project.R
\name{create_folders_oes_project}
\alias{create_folders_oes_project}
\title{Creates folders for an OES project}
\usage{
create_folders_oes_project()
}
\description{
Creates folders for an OES project
}
