% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_render_template_to_project.R
\name{data_extraction_render_template_to_project}
\alias{data_extraction_render_template_to_project}
\title{Places a copy of the data extraction render template into a project folder called scripts}
\usage{
data_extraction_render_template_to_project()
}
\value{
True or false
}
\description{
Places a copy of the data extraction render template into a project folder called scripts
}
