% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_with_sampling.R
\name{datatable_with_sampling}
\alias{datatable_with_sampling}
\title{Displays data as a datatable and samples if the dataframe is large}
\usage{
datatable_with_sampling(in_data, in_max_rows)
}
\arguments{
\item{in_data}{Dataframe to display}

\item{in_max_rows}{Maximum number of rows}
}
\value{
Datatable with with maximum number of rows
}
\description{
Displays data as a datatable and samples if the dataframe is large
}
