% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_unique.R
\name{filter_unique}
\alias{filter_unique}
\title{Filters out rows with NA values for specific columns}
\usage{
filter_unique(in_data, ...)
}
\arguments{
\item{in_data}{The dataframe you wish to remove rows from.}

\item{...}{Column(s) that determine whether a row should be removed.}
}
\value{
The dataframe with rows removed that had NAs in specified columns.
}
\description{
Filters out rows with NA values for specific columns
}
