% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Calculate the mode of a vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{The vector over which to calculateS the mode.}
}
\value{
The mode of the vector.
}
\description{
Calculate the mode of a vector
}
