% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_breaks.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Integer breaks}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{The number of breaks}

\item{...}{Other parameters passed to the scales::pretty function}
}
\value{
Integer breaks for ggplot
}
\description{
Integer breaks
}
