% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_date.R
\name{nice_date}
\alias{nice_date}
\title{Returns a nice looking date for text}
\usage{
nice_date(in_date, abbr = FALSE)
}
\arguments{
\item{in_date}{Any date value}

\item{abbr}{Boolean whether or not to abbreviate month name}
}
\value{
Text string of the date
}
\description{
Returns a nice looking date for text
}
\examples{
Enter as_date('2021-01-01'), returns "January 1, 2021"
Enter as_date('2021-01-01',abbr = TRUE), returns "Jan 1, 2021"
}
