% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_month_year.R
\name{nice_month_year}
\alias{nice_month_year}
\title{Returns a nice looking month and year for text}
\usage{
nice_month_year(in_date)
}
\arguments{
\item{in_date}{Any date value}
}
\value{
Text string of the month and year
}
\description{
Returns a nice looking month and year for text
}
\examples{
Enter nice_month_year('2021-01-01'), returns "January 2021"
}
