% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template_to_project.R
\name{render_template_to_project}
\alias{render_template_to_project}
\title{Places a copy of the render template into a project folder called scripts}
\usage{
render_template_to_project()
}
\value{
True or false
}
\description{
Places a copy of the render template into a project folder called scripts
}
