% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_folders.R
\name{set_up_folders}
\alias{set_up_folders}
\title{Creates a standard folder structure for project}
\usage{
set_up_folders()
}
\value{
Nothing if the folders do not exist and warning if they exist
}
\description{
Creates a standard folder structure for project
}
