% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_to_use_words.R
\name{when_to_use_words}
\alias{when_to_use_words}
\title{Converts number into text}
\usage{
when_to_use_words(in_number, sentence_begin = FALSE)
}
\arguments{
\item{in_number}{Number to transform into text}

\item{sentence_begin}{Logical for whether this is the start of a sentence}
}
\value{
String of the English word or number for text
}
\description{
Converts number into text
}
