% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_template_to_project.R
\name{word_template_to_project}
\alias{word_template_to_project}
\title{Places a copy of the word template into a project folder called templates and DHS logo and ADA into images}
\usage{
word_template_to_project()
}
\value{
True or false
}
\description{
Places a copy of the word template into a project folder called templates and DHS logo and ADA into images
}
